+++ unirest-mocks/src/main/java/kong/unirest/MockClient.java
@@ -41,6 +41,7 @@ import java.util.stream.Stream;
 public class MockClient implements Client, AsyncClient {
     private final Supplier<Config> config;
     private List<Routes> routes = new ArrayList<>();
+    private Invocation defaultResponse;
 
     public MockClient(Supplier<Config> config){
         this.config = config;
@@ -107,6 +108,7 @@ public class MockClient implements Client, AsyncClient {
     private Routes createNewPath(HttpRequest request) {
         Routes p = new Routes(request);
         routes.add(p);
+        p.addInvoke(defaultResponse);
         return p;
     }
 
@@ -176,5 +178,21 @@ public class MockClient implements Client, AsyncClient {
      */
     public void verifyAll() {
         routes.forEach(Routes::verifyAll);
+    }
+
+    /**
+     * Reset all expectations
+     */
+    public void reset() {
+        routes.clear();
+        defaultResponse = null;
+    }
+
+    /**
+     * return this status for any request that doesn't match a expectation
+     */
+    public ExpectedResponse defaultResponse() {
+        this.defaultResponse = new Invocation();
+        return this.defaultResponse;
     }
 }